Use static assets instead of compiling assets in production.
In order to compile an asset, we need to list it in config/application.rb
(config.assets.precompile) - unless we're listing it in an already-precompiled
file like application.css and application.js. Because we didn't fully understand
this, and because we were having problems with serving assets (possibly we hadnt
set config.assets.serve_static_assets to true) we turned on compilation of
assets in production. This was bad of course, but it worked. Well, with
backbone, it stopped working in a confusing way, so I undid it all.

This is a step up for now, since we're not compiling assets in production.
However, we are still serving assets from the Rails app, which is trivial to
move to nginx, which we should do next. I'm just waiting for Allen to do that,
since he did the  first bit of nginx.