Since dieter doesn't support nested assets yet, move things around so we don't need them.
Dieter doesn't support nested assets (that is, if a CSS file references an
image, it wont convert it to an asset). That means that static files don't get
the support that dieter provides - unique md5s, caching, joining, compressing,
etc. This sucks, but it's fine for now - not too hard to add later I think.