updated typefaces to 'PT' family
PT is pretty great. lots of variations in the family, everything
matches and flows nicely (including mono, which is rare) and it
simplifies a lot of things. i was hoping for more weight variations but
the lack of options is weight may actually pay off by simplifying the
number of choices and strengthening consistency through the app's type.