reset font size for button mixin
my original thoery for this was that in the increase font size would
invoke more emphasis; not needed in every case but i thought i was use
it more often than not. i was wrong. i found myself reseting it back to
normal size to fit inline with other elements. im undoing the size
increased and leaving it to individual elements to be changed if need-be