Deal with scrolling of the container viewport by browser search.
The element scroll handler finds out how far the scroll bas has been
moved by the search and works out which container should be selected for
that amount of scroll. It then selects that container and lets the
container selection code take care of animating the scroll.

The container selection code disables the scroll handler while it's
animating the scroll to the new position.