Reduce memory footprint of action logs
This is very much optimised for the happy path, assuming that people
won't actually jump around containers much and that they don't expand
successful actions much.

1) Ignore pusher messages for all but the current container
2) Drop output for any actions that are minimised when switching away
   from a container
3) Retrieve output for expanded action logs when switching to a
   container
4) Retrieve output whenever an action log is failed

The last one of these ensures that search always works for failed
actions in other containers since they'll always be expanded and always
have their output fetched.

The downside is that if you're looking at a container you see the
spinner when the action is marked as a failure, I consider this a UX
bug.