Gutted out the old invite stuff.
We don't use this anywhere, so we shouldn't have the API calls and routes
anymore.

(The bit in not_run_message.hamlc seems to have been dead code)

Also got rid of search-for-user -- Paul says that's an unindexed search
and it seems better to get that kind of thing from Github anyway.