Cordoned off all the invitations stuff.
Previously I was acting as if Circle users and Github users had the
same associated data. You may be surprised to learn this is not the
case ! I noticed that I had spuriously changed @github_id to @id
in User#gravatar_url and then had a very large epiphany.

I also renamed a lot of things to be more clear whether they refer
to the first green build invite form, and whether they refer to
github users or circle users.