More robust condition for getting Inivtations; moved logic to view.
This makes first_green_build_invitations lazily computed, and only
requests it in the view if display_first_green_build_invitations is
true.

This is more robust than what we had previously -- before the condition
for creating an Invitations included checking if there was an active project;
this is bad since it created the object and called github_users when
people were just on the project page. Now the view for build pages
checks the condition directly and then calls first_green_build_invitations,
a lazily-computed observable, if and only if the user is on a project page.