Don't show the loading section if we 404 on the /users API call.
This sets a an 'error' callback for the $.ajax request to /users that
forces the invitation form to redraw when the request fails.

When the invitation form redraws and sees that github_users and
github_users_not_following are null and loading_github_users is false,
it doesn't draw anything.

Partially fixes #2070.