Embed the page-number as a query parameter in the URL
I don't know of a way to determine how many builds are in the overall
view on the dashboard without querying for counts for every project
covered by a plan. As a result there isn't a UX for displaying how many
pages there are.

Instead I've chosen to add an obvious query parameter to the URL, this
gives people an easy way to skip around large build lists without
requiring specific UI (our customers are devs!).

There's a similar issue with disabling the "previous" button when you
reach the end of the build list. I've gone with disabling the button
when the number of builds is less than the builds-per-page. If the
number of builds divides evenly by builds-per-page then the back button
will take you to an empty list.