Mark actions as having output even though we don't append the output.
The bug was caused because actions were only marked as having output
when the action finished and a message was sent on the "updateAction"
pusher channel.

This change marks has_output true for an action if the frontend ever
sees an "appendAction" message for the action.

Fixes #2480 at the expense of duplicate log messages. Instead of not
seeing the previous messages now you are very likely to see duplicates
when you switch containers with the build running.

This is a hack wrapped up in a bodge, trading out one problem for
another.