Subscribe to observables that impact the height of the page.
Stops using ResizeSensor, it caused a lot of layout invalidation and
expensive paint operations as a result.

This new method is more complex than being notified whenever the page
changes size. It requires us to explicitly tell the build object what
observables impact the page height. It's an extra thing we need to keep
in sync with other code.