Flip observable subscription tracking on its head.
ActionLogs now have functions to subscribe/unsubscribe a watcher. When a
watcher is subscribed its "subscription_callback" function will be
invoked each time one of the ActionLog's observables change.

Subscriptions are automatically disposed when the container is switched
and new subscriptions are set up for the actions in the now-selected
container.

Cleaning an ActionLog disposes of any active Knockout subscriptions.