Use actual github avatar for user
Github user info contains an avatar_url which provides the avatar for
the user regardless of whether the avatar was uploaded directly to
github, comes from gravatar, or is a default provided by github.  This
uses that, so avatars uploaded to github are supported.

Currently only the logged in user; similar still needs to be done for
other places where github avatars show up