Mention badge generator in status badge docs
Making clear how completely trivial it is to do might increase usage.

Also there was a support request that had to do with an improperly
created badge (though that was a token issue so might not have been
prevented by this.)

The one thing that's slightly odd about this is that the docs
prominently mention the shield style while it was decided not to include
that in the settings page.  On the one hand the inconsistency feels a
bit annoying, but on the other hand it seems perfect to make clear that
people who care can use shield but make lazy/indecisive people use
badge.