Make popovers work.
This is a really awful hack :(
Overflow of the settings page content is set to "visible" when the
CodeDeploy settings are mounted into the DOM and set back to "auto" just
before they are unmounted.

Also add a bit more padding to stop the text input text from flowing
under the popover icon.