Logic part of org invites frontend
This isn't ready yet; it still needs the visual aspects.  Putting this
out there now in case there's significant feedback, while I start
fighting with CSS.

What this does is replace the invite-your-teammates-by-sending-this-link
modal with a regular page from which you can select your orgs to and use
the component from the first-green-build to invite them as easily as
checking a box if they have public github emails.

TODO:
- add CSS so that the list of orgs is a nice list on the left like the
  org list when adding projects
- refactor the CSS on the component stolen/refactored from
  first-green-build so that it still makes sense now that it's not
  necessarily first-green-build