Update junit docs
Our modification to `rspec_junit_formatter` got merged upstream so the
official version can be used now.

I also noticed that the description on the experimental settings page
was rather out of date; updated it mainly by making it shorter and
pointing to the docs so there are fewer places that need to be updated
now.