Add nginx to serve prod.localhost over SSL
We discovered a serious vulnerability related to our prod.circlehost
stuff - details here: https://github.com/circleci/circle/pull/5329

The solution is to use SSL during local dev too and to update the
backend to only work when using SSL - see the PR above.

This PR adds nginx configs and procfile config to make this work. We'll
need to merge the PR above and get it to prod before this will
work-work, so hold off on merging.