Fix cancel button icon alignment in Firefox.
Firefox currently doesn't support treating buttons as flexbox
containers, so putting the 'img' icon inside of the button didn't
have the desired effect. This moves the icon out of the button and
turns off the attempt to use the button as a flexbox container.

For more details on the flexbox bug, see: https://github.com/philipwalton/flexbugs#9-some-html-elements-cant-be-flex-containers