Add opt-in banner for ui-v2
We'll start to roll out ui-v2 by popping a banner that lets people
enable it. Under the covers we manage this with two feature flags - one
in launch darkly for the banner, and the other in a cookie. The cookie
feature flag is set or unset by the user using a new opt-in banner.

We'll use this pattern in the future, but I've made the opt-in stuff
very ui-v2 specific - this is intentional and intended to be refactored
in the future.