Change mechanism used to enable elevio/intercom
Previously, elevio/intercom would be enabled by adding a class to the
DOM node corresponding to widget we wanted to enable. This is
problematic because the DOM node might not exist at the time when the
class is added.

In this commit, elevio/intercom is enabled by adding a class to document.body