Add a banner reminding people the ios beta is ending
Add it to all builds on projects where "osx" is enabled.

This is pretty wonky - we're re-using a class that doesn't make sense,
we're adding current project to the main app "let" - but we need to get
this up asap and will be removing it within a couple weeks.

There's a CSS issue with the banner - hoping to get it fixed before we
push but worst case we can fix in prod.