Add ability to skip dispatching links
We need the ability to not dispatch some links internally so that we can
have some links render the new static site. Add the ability to add a
class that does this and add that class to a bunch of links in the
footer and on the landing and pricing pages.