Stop using no-local-state, disable instrumenting.
no-local-state doesn't work in 0.9.0, and we don't really need it
anymore anyhow.

Instrumenting breaks the app in this setup, and we don't use it (plus it
may not even work anymore). Commenting it out for now; we'll have to
decide whether to make it work or rip out the whole thing.