Upgrade cljs-time
This one's a little dangerous, because, according to the Readme:

> Note: version 0.3.* follows the API of clj-time 0.9.0.

(https://github.com/andrewmcveigh/cljs-time)

However, this may not mean there are any breaking changes to existing
APIs. Either way, the old version doesn't work with new ClojureScript,
and we'll need up upgrade eventually, so let's give it a shot.