Fix "doubled" test output
We're not exactly doubling the test output. Instead we're setting the
trailing output to the same value as the non-trailing output, which
means we display the non-trailing output twice and the trailing output
never. For instance, under "Starting the build", the first build step:

```
Installing inference daemon...
Inference daemon installed.
Starting inference daemon...
Inference daemon started.
```

the first three lines are normal output and the last is trailing output.
We're printing the first three lines twice and not printing the last
line.

The thing that keeps me from being confident in this fix is that I
actually have no idea what trailing output is.