Remove trailing output concept
Based on a reading of the code, it seems like all trailing output does
is pull out the last line separately. I have no idea why we do that, but
just conjing the trailing html and normal html together seems to work
well. The :converted-message key on the converters-state was only used
by Om, not but the coffeescript, so it's safe to remove as well.