Prevent all forms from submitting (by default)
Or rather, prevent buttons that would submit forms from submitting
forms. We want to preventDefault() at the click event, which prevents
both built-in validation and submission. Calling preventDefault() on the
submit event wouldn't stop the browser's built-in validation. (We could
do something that way, but this best preserves the current behavior.)