CIRCLE-283 Added rebuild button on dashboard
**This was written as part of an interview**

Currently reusing event from build page, which causes app to navigate
to the new build page.

Also refactored styles to be componentier. According to @peeja we can't
actually use components on this page until the full page is
componetized, but he wanted us to make the styles match the new format,
but just using class-names instead.