Remove apply-app-id-hack
The HTML we get from the server has a div#app where the app is meant to
render. We've been changing the id of that element to #om-app, as the
page loads, apparently for compatibility with the Knockout app we no
longer use.

Today, that's pure cruft, so stop doing that.

The div which renders just inside that div (the root component's DOM
node) was already #app, but we don't appear to use that id for anything,
so remove that id to avoid a clash.

While we're at it, when the :navigation-point hasn't been set yet,
rather than render an empty div, just render nil. The code is simpler
and it accomplishes the same thing.