:default-email should only be expected in identify calls
we added this to analytics, and initially I added it to our general
purpose UserEvent schema. Even though identify was the only call
currently using this, it was still wrong, so I moved it to an
IdentifyEvent schema.

We may want to remove the UserEvent schema in the future if not used,
but thats a problem for another day