Size main content even more correctly
This solves the horizontal overflow problem correctly, while also
solving a vertical scrolling problem. Before, the outer navigation
"chrome" was 70px taller than the viewport, which meant that it itself
could scroll in addition to the page content within it. Now it's sized
exactly to the viewport, so only the .app-dominant scrolls.