Track clicks of "Add Project" in empty state
Right now the "Add Project" link in the top-right of the projects page
is tracked as a `add-project-clicked` event. This modifies that event to
include an `is-empty-state` property with a value of `false`. A new
`add-project-clicked` event is added with an `is-empty-state` property
with a `true` value for clicks on the link in the empty state.