fixed some analytics events that were not view agnostic
our analytics events are meant to be view agnostic, to make it easier to
do rollups of all of the <x> that happened. especially since it is very
easy to do a group by in either sql or in amplitude on the event data
and see the more granular view, but its more difficult to aggregate
against multiple evnts (especially if you don't know that there are
other events you should be aggregating against).