nav-point->page is a constant again
This became a function because we can't call `config/enterprise?` early.
However, making it a function meant we were recreating some of the
component functions on every render. That meant that Om/React considered
every render to be rendering a different React "class", which caused
React to throw out the old component and mount a new one on every
render, rebuilding the entire DOM of the page.

Now, we again compute everything up front, so React can see each
rerender of a given page as the same component, and not remount it.