CIRCLE-1149 refesh user list (#1493)
When a user is removed from an org on GitHub, we need to remove them from the projects they follow of that org (unless OSS). Instead of waiting (a week?) for the user refresh cron job to do it for them, we need to allow the admins of an org to manually kick-off a refresh of all of the users. So we're giving them a button on the /team page to do so.