NUX Bootstrapped Experience Test
This test includes changing the empty state of the dashboard to fetch a
list of the most recent projects the user has interacted with on their
vcs provider and suggesting that user follows those projects on CCI.
This test is limited to github users only.

As part of this we updated the feature ns to be more testable

to do this I made feature/enabled? ^:dyanmic so that we can mock out its
return values in our tests.

I re-added feature/ab-test-treatment so that we can get the value for a
treatement (this takes an optional user argument in case your tests are
dependent on the user's state)

I made feature/ab-test-treatments more testable by using DI so that the
function is not dependent on other functions.