Made it so that users can filter to "their" builds only
Duplicated the Mine/All on the branch picker for this. Note that 'all' is
used to maintain consistency with branches and projects, but it is more
correct to filter builds on the back end using 'mine' (as is reflected
in the actual api call).

This commit sets up one approach to applying CSS to this new format, but
the changes to the CSS and class names should NOT be used, as they have a
high potential to introduce future formatting issues.

Note that users cannot filter builds on individual branches, as this
information is not indexed in the backend. A tooltip / disabled text
prevent them from attempting to do this.