Update :app/current-user remote reads after #1608
This should have been part of #1608. This updates :app/current-user's
remote-read to work the way Compassus now expects.

It also adds back a `read-remote` default implementation, but rather
than default to reading from the remote like we used to, this one just
throws a useful error message prompting the creator of a new key to
define its remote behavior. Asking for explicit direction here seems
better than making an implicit assumption that a new key will want to be
read from the remote.