Remove close button from user settings
It doesn't work correctly as implemented: if you're on `/account`, it
takes you to `/`, which isn't supposed to be an inner URL at all, and if
you're on `/account/heroku` (for instance), it takes you to `/account`,
which doesn't make much sense.

There isn't really a great place for this to take you to. Also,
`aside.less` says:

```
// [...] the close menu should be removed on completion of v2 anyways.
```

so apparently we were already expecting to lose these a while ago.