Improved the NUX Bootstrap Experience (#1625)
We have been seeing in the data a very low average projects shown for
the NUX Boostrap Experience. This improvement adds all currently
building projects on circleci that the user has permission to follow.

This makes the experience slow, as we do many round sequential trips to
the backend to fetch all this data. However we expect this to not be too
much of a burden as it is the users first time logging in and they dont
have anywhere else to go.

As part of this I have also improved the tracking, such that
nux-bootrap-impression will now have all the data we care about (such as
how many projects we are showing).

As part of this I also added map-vals to a map utils ns.

Also - dyanmic columning based on # of projects returned, and ability to see a
demo + event when people click on it.