added relevant build information to the elevio user for support (#1630)
Currently, when a user fills out a zendesk ticket via the elevio widget
on the website, we don't pass support much useful information. This
means that a lot of time support is flying blind as to what build,
project, or org the user is talking about.

This PR should give support some more information by passing
last-build-viewed and is-public-beta to support so they have a better
idea of what the user is talking about and can more easily service the
request.