Fix hover state on "link" by providing :href
I know this looks awful, but we do it because it *is* awful. This isn't
a real link, and it shouldn't be marked up as an `<a>`. That's what
we've done, though, and to get the correct styles it needs to have an
`href` attribute; otherwise it's an old-school anchor, not a "link".

In these cases, we've set a precedent of using `"javascript:void(0)"`
because it works, it's easy to remove later, and it's extremely obvious
that it's a hack.