Route internally to all relative links.
This change is explicitly made so that external links to docs / the
outer static site will break if they are not abosolute urls. All
relative paths will navigate internally, all absolute paths will be
external and trigger a hard refresh. On enterprise, we need the static
/ docs / outer links to be absolute so they are still navigable when
the host of the app is not circleci.com.