fix the elevio support information so that support actually gets it (#1632)
In my first iteration of this, I put all the stateful properties we were
trying to pass to support in the top level of the _elevio.user object.
It turns out for custom traits we need to store them in the
_elevio.user.traits map, so I fixed this.

Secondly, you cannot have nested maps in the _elevio.user.traits object.
Otherwise they just show up at "[object Object]". So, I recursively
flatten any map passed in and create a key which is the combination of
all keys passed in, underscore delimited. This way we can pass objects
and preserve the information of their context when passing values to
support.