Nux Bootstrap 4.0 - show all projects github and bitbucket (#1645)
After our learnings from 3.0 - that we do not have an accurate proxy for
user engagement and therefore showing all projects and letting the user
filter themselves increases the user's ability to follow projects they
care about and ultimately increases engagement.

Since in 3.0 we were showing all projects the user has the ability to
follow, but not all projects the user had the ability to add, we saw a
dramatic increase in follows but not a dramatic increase in adds (even
though a moderate one).

This treatment is going to show all projects a user has access to either
follow or build, across our two vcs providers (github and bitbucket).
Our working theory is that by showing all projects the user will choose
those most important to them, follow/build them, and will in the end be
more engaged.

One Caveat: There is no way to differentiate between GH and BB currently
in the treatment. So if the user is part of both a github and
bitbucket org of the same name, they may not know which org we are
referencing. We can fix this later if we think it is an actual problem.