[CIRCLE-3150] Add duplicate frontend event for backend signup (#1757)
**Rationale**
Some of the Segment Integrations that Marketing would like to use can only hook into frontend events to perform their functions. This is problematic because the event we're most interested in using the integrations with is `signup`, which is a backend event.

This PR works around this problem by firing a frontend event named `signup-event-sent` whenever a user logs in for the first time, which is generally the first frontend interaction that can occur after `signup`.

**Considerations**

This current implementation pulls the user's `sign-in-count` from the `renderContext`, which is populated from the server, meaning we can access it almost immediately without waiting for any other requests. 

After sending the event, we also set a localStorage flag to ensure the event doesn't get fired again from the same client. Having a per-client localStorage flag should be sufficient to cover the large majority of use cases because for this code to get called, `sign-in-count` needs to equal 1, and if a user accesses his/her account on more than one device, this condition will not be met for all but the initial device.

The only edge cases that I can think of where this can break down are: 
- If a user clears localStorage during their very first session (not much we can do about that)
- If another user signs up for CircleCI on the same client as a previous new user (there are ways we could address this, but it'd add complexity for very little gain, IMO)

Also note that there will inevitably be a rather significant discrepancy (10-15%) between the number of `signup-event-sent` vs `signup` events due to issues inherent to frontend events, like adblockers/tracking protection extensions, so these edge cases will likely only ever amount to a rounding error at worst.

**Ticket:** https://circleci.atlassian.net/browse/CIRCLE-3150