Fix frontend.utils macro requires
This moves all `:require-macro` forms for the `frontend.utils` macro
namespace to instead be a part of `:refer-macros` in the `:require`
form. For example,

`(:require-macros [frontend.utils :refer [html]])`

becomes

`(:require [frontend.utils :refer-macros [html]])`

The macros in the frontend.utils macro namespace generate code that has
references to vars in the frontend.utils cljs namespace, so the
`frontend.utils` namespace needs to be be required separately if the macro
ns is required via `:require-macros`. If someone mistakenly does not add
the cljs require, the code may break. By combining the require for the
cljs ns and the macro ns, we can ensure that the macros are never
required without the matching cljs namespace.