Attempt to fix bug where build timer not counting
This change contains a potential fix to the big where the build timer is
not counting. Since this cannot be reproduced in the development
frontend (running against both the dev or prod backend), this fix is
being tested in production under the feature flag :build-timer-fix.

If the fix is successful, a commit removing the feature flags will be necessary.