Fix container pill navigation (partial fix)
In the ever frustrating tale of the container pills navigation, there
are two current bugs (listed in order of priority):

It takes 2 clicks to navigate to a new container tab Every time we
navigate containers it refreshes the entire build page The first thing I
did here is to make all navigation happen via href. This gives us a
single point of entry for container + build navigation, and solves a lot
of issues about having different paths to the same result, which
resulted in having many solutions for the same problem.

That fixed the first issue. The second issue was only occurring on pages
where the tab had a dash in it, like the default tab queue-placeholder.
It turns out the regex for parsing the build url was not accounting for
any tabs not in a predefined list. This is very brittle so I removed it
and instead added a regex for any combination of alpha-numeric
characters and dashes. This should be more robust in the future.

That solved the second issue.

This still leaves an issue where clicking a non zero tab causes the
containers to jump from current-tab to zero then to new-tab, which is
not ideal. Also, the scrolling animation is terrible. However this is an
improvement on the previous version, and therefore we are going to merge
this as I debug that issue on a new branch.