Don't clear other `:current-*-data` on nav either
We shouldn't need to. There's a risk we have stale project or org data
temporarily and show something odd until the right one loads, but
there's no good way around that for now and it'll go away when this data
fetching moves to Om Next. (Stale build data would be more obvious, but
we're able to filter that out of the state by matching the build data
against the build we expect to find there.)