Add logging to debug infinite project spinner bug
This bug has been a strange combination of things that makes it almost
impossible to track down. So far I have confirmed:
1. The infinite spinner is caused because the :projects key is nil, but
the projects api call is clearly made, and returns a non-nil response.
2. It only happens for specific users or users within specific orgs,
but happens deterministically for them.
3. It only happens with advanced compilation turned on. Appending
`om-build-id=whitespace` to circleci.com makes the issue go away.
4. Using advanced compilation and prod.circlehost and impersonating one
of the affected users, the issue also seems to go away.

Given those conditions, the only next step I can think of is to deploy
the logging to production. This will be removed once the issue is
solved.