Fix rerun read delay
The rerun read and the initial read had become interchangeable, so both
were using the initial read. They did the same thing, except the rerun
read was on a delay. This delays a read after a rerun no matter what it
is, and then uses the same cond branch to implement the read.