Refactor the org centric UI to have a better architecture
There were a couple things that this commit is meant to address:

standardize the way we redirect from old, non org centric routes
standardize the way we show 404 pages when a user does not have access
to a page make all org-centric routing use hrefs (give a single point of
entry to each page) First we split the selected org into 2 different
concepts: the org the user navigated to and the last org the user
visited. The org the user navigated is simply parsed from the route. The
last org the user visited is stored in local storage and is used for
redirection if they land on an old, non org-specific route (such as
/dashboard).

We use the org stored in local storage only for the purpose of
redirecting. This org is set when a user navigates to a org which is in
that user's list of orgs (the list of orgs they are a part of in their
vcs provider). This means that when site admins (like support) visit
customer's pages, that org will not get set in their local storage.

We use the org parsed out of the route for everything else. This way we
set the route data (which can be inputted by a user) in a transitive
state, so we never redirect a user to a page that will 404 (which is the
epitome of shitty UX).

We show 404s by using the "core" api route on each page to determine
whether or not the currently auth'd user has permission to view the page
they are trying to view. If they do not, then we redirect to a 404 page.

A few pages (like build-insights) still use user-centric routes (give me
all insights for this user). On these pages we fetch the org's plan to
ensure that this user has access to that page.

This PR does not touch Om Next, since Nate is working on 404s for those
pages.