Set-of-keys resolvers get all the ASTs as a map
This makes it easier to handle multiple keys with a single resolver.

Note that the resolver doesn't really get all of the ASTs yet, just the
one it's working on, because we call the resolver for each key rather
than call it just once. This API means that we can call the resolver
just once in the future, though.