fix 404 bug on the project-settings#env-var page
We currently use the org-settings endpoint to get a list of projects for an org
for the bulk import var feature on env-vars. However, you can have read and write
access to projects which belong to orgs you are not a member of (not sure how but
I did confirm that a thing).