moved :selected-org under the navigation-data
its not really part of the user, its part of the navigation. and either
way, it was causing issues since we use whether or not the
state/user-path is nil? as whether or not to show the login button on
the 404 page (the only way to log in on prod.circlehost currently)