made the ellipsis work with the breadcrumbs
I had to remove the 2.0 platform indicator from the breadcrumbs and move
it to the actions, which is probably where it should have been anyways.
I also had to make the breadcrumbs display: inline rather than flexy to
get the ellipsis to work.

with these changes everything overflows as expected.