quick fix to make the cancel modal a little more forboding
our current cancel modal really emphasizes you telling us why you are
cancelling, it does not emphasize what you are going to lose by
cancelling. this makes a change so that the premium features you get as
a paid member will go away immediately is front and center.

we have also added "Build History" as the top feature you will lose
because we believe this is one of the main things people care about.