Fix some Closure warnings
Loom uses the name `in` as a record property, which then becomes a JS object
property. In EC3, `in` (along with other keywords and reserved words) aren't
allowed to be property names. But EC3 is ancient now: modern browsers use
EC5 (plus various bits from EC6). We just have to tell the Closure compiler that
we're expecting to have EC5 code, and it'll stop warning us.